# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação Pedidos de viagens",
        "description": "Swagger Pedidos de viagens",
        "contact": {
            "email": "gustavo.caldeira1@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "description": "List all orders",
                "operationId": "325fd3d55bbe9bb39645269928fa4983",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit records per page",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "9"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna notícias da agencia sebrae",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaginatedOrderResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Exception error message",
                                        "value": {
                                            "message": "Example error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Exception error message",
                                        "value": {
                                            "message": "Example error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create order",
                "description": "Create order",
                "operationId": "445bab4f7dc9ea6fcd54f286c0209478",
                "requestBody": {
                    "description": "Requested data to create a order",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "message": {
                                        "summary": "Success message",
                                        "value": {
                                            "message": "Saved successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Exception error message",
                                        "value": {
                                            "message": "Example error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseException"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Exception error message",
                                        "value": {
                                            "message": "Example error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "description": "Return all data from a order by ID",
                "operationId": "f088d34a2737782215f435f854e31de5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return all data from a order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Exception error message",
                                        "value": {
                                            "message": "Example error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Error message",
                                        "value": {
                                            "message": "Order not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Exception error message",
                                        "value": {
                                            "message": "Example error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/orders/{id}/status": {
            "patch": {
                "tags": [
                    "Order"
                ],
                "summary": "Update order status",
                "description": "Update order status",
                "operationId": "579bcf347ab0f4a9c09499b54d03176f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Needed data to update order status",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Exception error message",
                                        "value": {
                                            "message": "Example error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Error message",
                                        "value": {
                                            "message": "Order not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpResponseException"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Exception error message",
                                        "value": {
                                            "message": "Example error message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "HttpResponseException": {
                "title": "HttpResponseException",
                "description": "Class HttpResponseException",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Main error message",
                        "type": "string",
                        "format": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "properties": {
                            "error_key": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Error message refering to error_key"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "HttpResponseException"
                }
            },
            "OrderStoreRequest": {
                "title": "OrderStoreRequest",
                "description": "Class OrderStoreRequest",
                "required": [
                    "interesses"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "User name",
                        "type": "string",
                        "format": "string",
                        "example": "João de Souza"
                    },
                    "destination": {
                        "title": "destination",
                        "description": "User destination",
                        "type": "string",
                        "format": "string",
                        "example": "Santo Antônio do Itambé"
                    },
                    "departure_date": {
                        "title": "departure_date",
                        "description": "Departure date",
                        "type": "string",
                        "format": "string",
                        "example": "2023-02-13"
                    },
                    "return_date": {
                        "title": "return_date",
                        "description": "Return date",
                        "type": "string",
                        "format": "string",
                        "example": "2023-02-14"
                    },
                    "status": {
                        "title": "status",
                        "description": "Request Status",
                        "type": "string",
                        "format": "string",
                        "default": "Requested",
                        "enum": [
                            "Requested",
                            "Approved",
                            "Canceled"
                        ],
                        "example": "Requested"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderStoreRequest"
                }
            },
            "OrderUpdateRequest": {
                "title": "OrderUpdateRequest",
                "description": "Class OrderUpdateRequest",
                "required": [
                    "interesses"
                ],
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Request Status",
                        "type": "string",
                        "format": "string",
                        "default": "Approved",
                        "enum": [
                            "Requested",
                            "Approved",
                            "Canceled"
                        ],
                        "example": "Approved"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderUpdateRequest"
                }
            },
            "OrderResource": {
                "title": "OrderResource",
                "description": "Class OrderResource",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "User name",
                        "type": "string",
                        "format": "string",
                        "example": "João de Souza"
                    },
                    "destination": {
                        "title": "destination",
                        "description": "User destination",
                        "type": "string",
                        "format": "string",
                        "example": "Santo Antônio do Itambé"
                    },
                    "departure_date": {
                        "title": "departure_date",
                        "description": "Departure date",
                        "type": "string",
                        "format": "string",
                        "example": "2023-02-13"
                    },
                    "return_date": {
                        "title": "return_date",
                        "description": "Return date",
                        "type": "string",
                        "format": "string",
                        "example": "2023-02-14"
                    },
                    "status": {
                        "title": "status",
                        "description": "Request Status",
                        "type": "string",
                        "format": "string",
                        "default": "REQUESTED",
                        "enum": [
                            "REQUESTED",
                            "APPROVED",
                            "CANCELED"
                        ],
                        "example": "REQUESTED"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderResource"
                }
            },
            "PaginatedOrderResource": {
                "title": "PaginatedOrderResource",
                "description": "Class PaginatedOrderResource",
                "properties": {
                    "currentPage": {
                        "title": "currentPage",
                        "description": "Página atual",
                        "format": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderResource"
                        }
                    },
                    "lastPage": {
                        "title": "lastPage",
                        "description": "Ultima página",
                        "format": "integer",
                        "example": 10
                    },
                    "total": {
                        "title": "total",
                        "description": "Total de páginas",
                        "format": "integer",
                        "example": 15
                    },
                    "perPage": {
                        "title": "perPage",
                        "description": "Ultima página",
                        "format": "integer",
                        "example": 10
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PaginatedOrderResource"
                }
            }
        },
        "securitySchemes": {
            "api_token": {
                "type": "apiKey",
                "description": "Please enter the authorization token.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Order",
            "description": "Endpoints to control travel orders"
        }
    ]
}